Rename EMC observable modules
