+++ acm/observables/emc/minkowski_module.py
