+++ acm/observables/emc/projected_tpcf_module.py
