+++ acm/observables/emc/recon_spectrum_module.py
