+++ acm/observables/emc/spectrum_module.py
