+++ acm/observables/emc/versus_module.py
