+++ acm/observables/emc/vide_module.py
